pyproject.toml
[tool.poetry]
name = "final"
version = "0.1.0"
description = ""
authors = ["Merve YILDIRIM <merve.yildirim@etu.univ-tours.fr>"]

[tool.poetry.dependencies]
python = "^3.09"
rich = "^12.3.0"
pyserde = {extras = ["all"], version = "^0.7.3"}
typer = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
